requirements.txt
Django==3.2
django-mptt==0.12.0
